+++ service/src/main/java/edu/eetac/dsa/flatmates/ListaCompraResource.java
@@ -5,6 +5,7 @@ import edu.eetac.dsa.flatmates.dao.GrupoDAOImpl;
 import edu.eetac.dsa.flatmates.dao.ListaCompraDAO;
 import edu.eetac.dsa.flatmates.dao.ListaCompraDAOImpl;
 import edu.eetac.dsa.flatmates.entity.ColeccionListaCompra;
+import edu.eetac.dsa.flatmates.entity.Grupo;
 import edu.eetac.dsa.flatmates.entity.GrupoUsuario;
 import edu.eetac.dsa.flatmates.entity.ListaCompra;
 
@@ -13,6 +14,7 @@ import javax.ws.rs.core.*;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.sql.SQLException;
+import java.util.List;
 
 /**
  * Created by Admin on 24/11/2015.
@@ -22,7 +24,7 @@ public class ListaCompraResource {
     @Context
     private SecurityContext securityContext;
 
-    @Path("/{id}")
+    @Path("/grupo/{id}")
     @POST
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(FlatmatesMediaType.FLATMATES_LISTA)
@@ -34,8 +36,12 @@ public class ListaCompraResource {
         String userid = securityContext.getUserPrincipal().getName();
         GrupoUsuario grupoUsuario = null;
         GrupoDAO grupoDAO = new GrupoDAOImpl();
+        Grupo grupo = null;
         try {
-            grupoUsuario = grupoDAO.getGrupoUserById(id);
+            grupo = grupoDAO.getGrupoById(id);
+            if (grupo == null)
+                throw new NotFoundException("Group with id=" + id + " not found");
+            grupoUsuario = grupoDAO.getGrupoUserById(id, userid);
             if(!grupoUsuario.getUserid().equals(userid))
                 throw new ForbiddenException("operation not allowed");
             listaCompra = listaCompraDAO.createLista(item, id);
@@ -45,6 +51,7 @@ public class ListaCompraResource {
         URI uri = new URI(uriInfo.getAbsolutePath().toString() + "/" + listaCompra.getId());
         return Response.created(uri).type(FlatmatesMediaType.FLATMATES_MENSAJE).entity(listaCompra).build();
     }
+    @Path("/grupo/{id}")
     @GET
     @Produces(FlatmatesMediaType.FLATMATES_LISTA_COLLECTION)
     public ColeccionListaCompra getListas(@PathParam("id") String id){
@@ -53,8 +60,12 @@ public class ListaCompraResource {
         String userid = securityContext.getUserPrincipal().getName();
         GrupoUsuario grupoUsuario = null;
         GrupoDAO grupoDAO = new GrupoDAOImpl();
+        Grupo grupo = null;
         try {
-            grupoUsuario = grupoDAO.getGrupoUserById(id);
+            grupo = grupoDAO.getGrupoById(id);
+            if (grupo == null)
+                throw new NotFoundException("Group with id=" + id + " not found");
+            grupoUsuario = grupoDAO.getGrupoUserById(id, userid);
             if(!grupoUsuario.getUserid().equals(userid))
                 throw new ForbiddenException("operation not allowed");
             listaCompraCollection = listaCompraDAO.getLista(id);
@@ -64,7 +75,7 @@ public class ListaCompraResource {
 
         return listaCompraCollection;
     }
-    @Path("/{idlista}")
+    @Path("/{idlista}/grupo/{id}")
     @GET
     @Produces(FlatmatesMediaType.FLATMATES_LISTA)
     public ListaCompra getLista(@PathParam("id") String id, @PathParam("idlista") String idlista){
@@ -73,8 +84,12 @@ public class ListaCompraResource {
         String userid = securityContext.getUserPrincipal().getName();
         GrupoUsuario grupoUsuario = null;
         GrupoDAO grupoDAO = new GrupoDAOImpl();
+        Grupo grupo = null;
         try {
-            grupoUsuario = grupoDAO.getGrupoUserById(id);
+            grupo = grupoDAO.getGrupoById(id);
+            if (grupo == null)
+                throw new NotFoundException("Group with id=" + id + " not found");
+            grupoUsuario = grupoDAO.getGrupoUserById(id, userid);
             if(!grupoUsuario.getUserid().equals(userid))
                 throw new ForbiddenException("operation not allowed");
             listaCompra = listaCompraDAO.getListaById(idlista);
@@ -86,24 +101,28 @@ public class ListaCompraResource {
         return listaCompra;
     }
 
-    @Path("/{idt}")
+    @Path("/{idt}/grupo/{id}")
     @PUT
     @Consumes(FlatmatesMediaType.FLATMATES_LISTA)
     @Produces(FlatmatesMediaType.FLATMATES_LISTA)
     public ListaCompra updateLista(@PathParam("id") String id, @PathParam("idt") String idt, ListaCompra listaCompra) {
-        if(listaCompra == null)
+        if(listaCompra.getId() == null||listaCompra.getItem()==null||listaCompra.getGrupoid()==null)
             throw new BadRequestException("entity is null");
-        if(!id.equals(listaCompra.getId()))
+        if(!idt.equals(listaCompra.getId()))
             throw new BadRequestException("path parameter id and entity parameter id doesn't match");
         String userid = securityContext.getUserPrincipal().getName();
         GrupoUsuario grupoUsuario = null;
         GrupoDAO grupoDAO = new GrupoDAOImpl();
+        Grupo grupo = null;
         ListaCompraDAO listaCompraDAO = new ListaCompraDAOImpl();
         try {
-            grupoUsuario = grupoDAO.getGrupoUserById(id);
+            grupo = grupoDAO.getGrupoById(id);
+            if (grupo == null)
+                throw new NotFoundException("Group with id=" + id + " not found");
+            grupoUsuario = grupoDAO.getGrupoUserById(id, userid);
             if(!grupoUsuario.getUserid().equals(userid))
                 throw new ForbiddenException("operation not allowed");
-            listaCompra = listaCompraDAO.updateLista(idt, listaCompra.getId());
+            listaCompra = listaCompraDAO.updateLista(listaCompra.getId(), listaCompra.getItem(), listaCompra.getGrupoid());
             if(listaCompra == null)
                 throw new NotFoundException("Item with id = "+id+" doesn't exist");
         } catch (SQLException e) {
@@ -111,15 +130,24 @@ public class ListaCompraResource {
         }
         return listaCompra;
     }
-    @Path("/{idt}")
+
+    @Path("/{idt}/grupo/{id}")
     @DELETE
     public void deleteLista(@PathParam("id") String idg, @PathParam("idt") String id) {
         String userid = securityContext.getUserPrincipal().getName();
         ListaCompraDAO listaCompraDAO = new ListaCompraDAOImpl();
         GrupoDAO grupoDAO = new GrupoDAOImpl();
         GrupoUsuario grupoUsuario = null;
+        Grupo grupo = null;
+        ListaCompra listaCompra = null;
         try {
-            grupoUsuario = grupoDAO.getGrupoUserById(idg);
+            grupo = grupoDAO.getGrupoById(idg);
+            if (grupo == null)
+                throw new NotFoundException("Group with id=" + idg + " not found");
+            listaCompra = listaCompraDAO.getListaById(id);
+            if (listaCompra == null)
+                throw new NotFoundException("Item with id=" + id + " not found");
+            grupoUsuario = grupoDAO.getGrupoUserById(idg, userid);
             if(!grupoUsuario.getUserid().equals(userid))
                 throw new ForbiddenException("operation not allowed");
             if(!listaCompraDAO.deleteLista(id))
@@ -128,5 +156,35 @@ public class ListaCompraResource {
             throw new InternalServerErrorException();
         }
     }
-
+    @Path("/{id}/grupo/{idgrupo}/hecho")
+    @POST
+    @Produces(FlatmatesMediaType.FLATMATES_LISTA)
+    public void hecho(@PathParam("id") String id, @PathParam("idgrupo") String idg)  {
+        String userid = securityContext.getUserPrincipal().getName();
+        GrupoDAO grupoDAO = new GrupoDAOImpl();
+        Grupo grupo = null;
+        ListaCompra listaCompra = null;
+        GrupoUsuario grupoUsuario = null;
+        ListaCompraDAO listaCompraDAO = new ListaCompraDAOImpl();
+        try{
+            grupo = grupoDAO.getGrupoById(idg);
+            if (grupo == null)
+            {
+                throw new NotFoundException("Group with id=" + idg + " not found");
+            }
+            grupoUsuario = grupoDAO.getGrupoUserById(idg, userid);
+            if(!grupoUsuario.getUserid().equals(userid))
+                throw new ForbiddenException("You aren't the admin");
+            listaCompra = listaCompraDAO.getListaById(id);
+            if (listaCompra == null)
+            {
+                throw new NotFoundException("Item with id = "+id+" doesn't exist");
+            }
+            if (listaCompra.isHecho())
+                throw new ForbiddenException("This item is did");
+            listaCompraDAO.HechoLista(id, idg);
+        } catch (SQLException e) {
+            throw new InternalServerErrorException();
+        }
+    }
 }
